refactor and clean tests
