+++ vllm/attention/backends/sink_rotations.py
@@ -1,26 +1,19 @@
 from typing import Callable, Optional
-from unittest.mock import MagicMock
 
 import torch
 
 
 class BackedUpSink:
-    def __init__(self):
+    def __init__(self) -> None:
         self.sink_key_cache: Optional[torch.Tensor] = None
         self.sink_blocks: Optional[torch.Tensor] = None
 
-    def register(self, sink_key_cache: torch.Tensor, sink_blocks: torch.Tensor):
+    def register(self, sink_key_cache: torch.Tensor, sink_blocks: torch.Tensor) -> None:
         """sink_blocks is flattened"""
         self.sink_key_cache = sink_key_cache
         self.sink_blocks = sink_blocks
 
-    # def __iter__(self):
-    #     for batch_i, (backup, blocks) in enumerate(
-    #         zip(self.sink_key_cache, self.sink_blocks)
-    #     ):
-    #         yield batch_i, backup, blocks
-
-    def __len__(self):
+    def __len__(self) -> int:
         return len(self.sink_blocks)
 
 
@@ -31,7 +24,7 @@ class SinkAttentionRotaryImpl(torch.nn.Module):
         sliding_window_size: int,
         num_kv_heads: int,
         head_size: int,
-    ):
+    ) -> None:
         super().__init__()
         self.sink_size = sink_size
         self.sliding_window_size = sliding_window_size
@@ -46,26 +39,21 @@ class SinkAttentionRotaryImpl(torch.nn.Module):
         self, key_cache: torch.Tensor, backed_up_sink: BackedUpSink
     ) -> None:
         key_cache[backed_up_sink.sink_blocks] = backed_up_sink.sink_key_cache
-        # for _, backup, blocks in backed_up_sink:
-        #     key_cache[blocks] = backup
 
     def process_decode_metadata(
-        self, attn_metadata, key_cache: torch.Tensor, rotary_emb: Callable, positions
+        self,
+        attn_metadata,
+        key_cache: torch.Tensor,
+        rotary_emb: Callable,
+        positions: torch.Tensor,
     ) -> BackedUpSink:
         decode_meta = attn_metadata.decode_metadata
         backed_up_sink = BackedUpSink()
 
         if self.sink_size > 0:
-            self._prepare_sink_rotation(decode_meta, key_cache, backed_up_sink, positions)
-            self._rotate_sinks(key_cache, rotary_emb, backed_up_sink, positions)
+            self._backup_sink(decode_meta, key_cache, backed_up_sink)
+            self._rotate_sinks(rotary_emb, key_cache, backed_up_sink, positions)
         return backed_up_sink
-
-    def _prepare_sink_rotation(
-        self, decode_meta, key_cache: torch.Tensor, backed_up_sink: BackedUpSink, positions: torch.Tensor
-    ):
-        """Prepare and return backup of sink positions for potential restoration."""
-        # for batch_i, batch_context_len in enumerate(decode_meta.context_lens):
-        self._backup_sink(decode_meta, key_cache, backed_up_sink)
 
     def _backup_sink(
         self,
@@ -73,50 +61,45 @@ class SinkAttentionRotaryImpl(torch.nn.Module):
         key_cache: torch.Tensor,
         backed_up_sink: BackedUpSink,
     ) -> None:
-        num_sinks_current = (
-            self.sink_size // key_cache.shape[-2]
-        )
+        num_sinks_current = self.sink_size // key_cache.shape[-2]
         sink_blocks = decode_meta.block_tables[:, :num_sinks_current].flatten()
-        # sink_key_cache = torch.index_select(key_cache, index=sink_blocks, dim=0)
         sink_key_cache = torch.index_select(key_cache, index=sink_blocks, dim=0)
         backed_up_sink.register(sink_key_cache.clone(), sink_blocks)
 
     def _rotate_sinks(
         self,
+        rotary_emb: Callable,
         key_cache: torch.Tensor,
-        rotary_emb: MagicMock,
         backed_up_sink: BackedUpSink,
         positions: torch.Tensor,
-    ):
+    ) -> None:
         # get rotations angles
         num_evictions = self._calculate_evictions(positions)
+        bs = positions.shape[0]
 
         rotation_positions = (self._dummy_rotations * num_evictions[:, None]).to(int)
 
         # rotate
-        bs = positions.shape[0]
-        sink_to_rotate = backed_up_sink.sink_key_cache.permute(0, 3, 1, 2, 4).reshape(bs, self.sink_size, -1)
-        # sink_to_rotate = self._format_key_cache_to_rotation(backed_up_sink.sink_key_cache)
-        dummy_query = self._dummy_query.repeat(sink_to_rotate.shape)  # torch.zeros_like(sink_to_rotate).to(key_cache.device)
+        sink_to_rotate = self._format_key_cache_to_rotation(
+            backed_up_sink.sink_key_cache, bs
+        )
+        dummy_query = self._dummy_query.repeat(sink_to_rotate.shape)
         _, rotated_sinks = rotary_emb(rotation_positions, dummy_query, sink_to_rotate)
-        # dupa = sink_to_rotate.min(dim=-1).values
         # Put correctly rotated sinks into the original position in the cache
-        key_cache[backed_up_sink.sink_blocks] = rotated_sinks.view(bs,
-                                                                   self.sink_size,
-                                                                   self.num_kv_heads,
-                                                                   self.head_size // 8,
-                                                                   8,
-                                                                   ).permute(0, 2, 3, 1, 4)
+        key_cache[backed_up_sink.sink_blocks] = self._format_rotated_to_key_cache(
+            rotated_sinks, bs
+        )
 
-    def _format_key_cache_to_rotation(self, x):
-        # in: bs,  num_kv_heads, self.head_size/8, 16, 8
-        return x.permute(3, 0, 1, 2, 4).reshape(self.sink_size, -1)
+    def _format_key_cache_to_rotation(self, x, bs) -> torch.Tensor:
+        return x.permute(0, 3, 1, 2, 4).reshape(bs, self.sink_size, -1)
 
-    def _calculate_evictions(self, positions: torch.Tensor):
-        p_i = positions
-        cs = self.cache_size
-        diff = p_i - cs
-        return torch.max(diff, self._cache_zeros)
+    def _format_rotated_to_key_cache(self, x, bs) -> torch.Tensor:
+        return x.view(
+            bs, self.sink_size, self.num_kv_heads, self.head_size // 8, 8
+        ).permute(0, 2, 3, 1, 4)
+
+    def _calculate_evictions(self, positions: torch.Tensor) -> torch.Tensor:
+        return torch.max(positions - self.cache_size, self._cache_zeros)
 
     # Method to set the device of cache_size_gpu
     def to(self, device):
